connection header fix
